fix: correct the clean behavior

+++ clean_my_mac.sh
@@ -1,42 +1,48 @@
 source "$CLEAN_MY_MAC/lib/colors.sh"
 source "$CLEAN_MY_MAC/lib/targets.sh"
-
-space_convert() {
-  local number=$(expr $1 + 0)
-  local UNITS=("K" "M" "G" "T")
-  for unit in ${UNITS[@]}; do
-    if [[ number > 1024 ]]; then
-      number=$(expr $number / 1024)
-    else
-      echo "$number$unit"
-      break
-    fi
-  done
-}
+source "$CLEAN_MY_MAC/lib/detect.sh"
 
 clean_my_mac() {
+  set -f # disable globbing
 
-  local total=0
+  CLEAN_TARGETS=$(targets)
+  TARGETS=( $(echo $CLEAN_TARGETS | tr ':' ' ') )
 
-  for target in ${targets[@]}; do
-    find ${target} >/dev/null 2>&1 || {
-      printf "${RED}Error: ${target} does not exist${NORMAL}"
-      continue
-    }
-    printf "${YELLOW}Estimating size of ${target} ... "
-    local space=$(du -shk ${target} | cut -f 1)
-    total=$(expr $total + space)
-    printf "$(space_convert $space).${NORMAL}\n"
-    printf "${BLUE}Removing ${target} ... "
-    rm -rf ${target} >/dev/null 2>&1 || {
-      printf "${RED}Error removing ${target}.${NORMAL}\n"
-    }
+  for target in ${TARGETS[@]}; do
+    set +f
+    if [[ "$(detect_shell)" = "zsh" ]]; then
+      if [ $(command find ${~target} 2>/dev/null | wc -l) -eq 0 ]; then
+        continue
+      fi
+      printf "${YELLOW}Estimating size of ${~target} ... "
+      space=$(command du -ch ${~target} | tail -n 1 | cut -f 1)
+
+      printf "$space.${NORMAL}\n"
+      printf "${BLUE}Removing ${~target} ... "
+      rm -rf ${~target} >/dev/null 2>&1 || {
+        printf "${RED}Error removing ${target}.${NORMAL}\n"
+      }
+    elif [[ "$(detect_shell)" = "bash" ]]; then
+      if [ $(command find ${target} 2>/dev/null | wc -l) -eq 0 ]; then
+        continue
+      fi
+      printf "${YELLOW}Estimating size of ${target} ... "
+      space=$(command du -ch ${target} | tail -n 1 | cut -f 1)
+
+      printf "$space.${NORMAL}\n"
+      printf "${BLUE}Removing $target ... "
+      rm -rf ${target} >/dev/null 2>&1 || {
+        printf "${RED}Error removing ${target}.${NORMAL}\n"
+      }
+    fi
+
     printf "${BLUE}done.${NROMAL}\n"
+    set -f
   done
 
-  printf "${GREEN}Clean up of total size: $(space_convert $total).${NORMAL}\n\n"
+  set +f
 
-  local SYSTEM_CACHES=("/tmp" "var/" "private")
+  local SYSTEM_CACHES="/tmp, /var, /private"
   printf "${YELLOW}If you wish to optimize your system caches in $SYSTEM_CACHES, reboot your Mac\n"
   printf "${YELLOW}Don't manually delete ANY files in there, it could crash your Mac\n"
   # DON'T EVER TRY MANUALLY DELETE SYSTEM FOLDERS, YOUR MAC MIGHT GET INTO TROUBLE!!
